Create a GuardianContentService and move tasks from MasterclassDataService into it. Add a new task for fetching content based on eventbrite reference type.
