+++ frontend/app/Global.scala
@@ -19,7 +19,8 @@ object Global extends WithFilters(CheckCacheHeadersFilter, CacheSensitiveCSRFFil
     MasterclassEventService.start()
 
     TouchpointBackend.All.foreach(_.start())
-    MasterclassDataService.contentTask.start()
+    GuardianContentService.masterclassContentTask.start()
+    GuardianContentService.eventbriteContentTask.start()
   }
 
   override def onHandlerNotFound(request: RequestHeader): Future[Result] = {
